add an item functionality done?

+++ app/forms/menu_item_form.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField
+from wtforms import StringField, IntegerField, FloatField
 from wtforms.validators import DataRequired, ValidationError
 from app.models import MenuItem
 
@@ -12,7 +12,7 @@ def menu_item_name_exist(form, field):
 
 class MenuItemForm(FlaskForm):
     name = StringField('name', validators=[DataRequired(), menu_item_name_exist])
-    price = IntegerField('price', validators=[DataRequired()])
+    price = FloatField('price', validators=[DataRequired()])
     description = StringField('description', validators=[DataRequired()])
     category = StringField('category', validators=[DataRequired()])
     photo_url = StringField('photo_url', validators=[DataRequired()])
